LPS-111561 Isolate Style Display selector
